+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Added
+
+- plugin state API
+- ability to change the color of lines on columns and captions
+
+## Changed
+
+- title size
+- improved adaptability of chart axes
+
 ## [0.5.0]
 
 ### Added